requirements.txt
opencv-python
pygame
requests
google-auth
google-cloud-aiplatform
